register data serializers with Forge fixes some netty DecoderExceptions
